==> 2a_woodTraits_explainDecay.Rmd <==
---
title: "Do wood traits explain decay?"
author: "Marissa Lee"
date: "12/2/2018"
output: github_document
---

```{r, include = F}
#chunk options
knitr::opts_chunk$set(echo=FALSE, message=FALSE, warning=FALSE)
```

Load libraries and functions
```{r loadlib}

#---------------------------#
#libraries
library(tidyverse) # ggplot2, dplyr, tidyr, readr, purrr
library(tidymodels) # broom
library(sjstats) #provides function eta_sq() to extract r2 for each predictor in a lm
library(grid) #plotting
library(gridExtra) # plotting
library(GGally) # bivariate plots
#devtools::install_github("cornwell-lab-unsw/litterfitter") #fit decay models to timeseries
library(litterfitter) #fit decay models to timeseries
library(lmtest) # for coxtest()

#library(vegan) #diversity metrics
#library(rioja) #WA-PLS
#library(knitr) #r markdown
#library(magrittr) #data formatting


#---------------------------#
#fxns
source('code/helper_fxns.R')
sourceDir('code')

```

Load decay data and wood trait data
```{r loaddata}
#---------------------------#
#data

# stem mass
initial_mass <- read_in_initial_mass()
harvest_mass <- LoadHarvestFiles()
pmr <- Calc_massRemaining(initial_mass, harvest_mass) # calculate percent mass remaining (pmr)
pmr_byStem <- AvgPMR_byStem(pmr) # average pmr for each stem and timepoint
stem.respVars <- list("time7", "time13", "time25", "time37","time59")

# decay fits
stemSamples <- load_stemSamples() # this function is in load_microbeData.R
#decayfits <- fit_all_curves(pmr, stemSamples) #this recalculates all the curve fits, uncomment if the data change; for each species+size
#write_csv(decayfits,"derived_data/decayfits.csv")
decayfits <- read_csv("derived_data/decayfits.csv")
indx <- unique(stemSamples[,c("code","Binomial","size")])
decayfits %>%
  left_join(indx) -> decayfits
code.respVars <- list("w.t50","alpha","beta","w.r2") # weibull params: alpha = shape param, beta = scale param

# wood traits
traits.code <- trait.means_byCode(stemSamples, fill.densitybark = TRUE) # averaged by code; if fill.densitybark == TRUE, then use small stem estimates to approximate large stem density and bark thickness
#trait.sds_byCode(stemSamples) # within code variation
#trait.n_byCode(stemSamples) # number of code samples that were aggregated
traits.stem <- trait.means_byStem(stemSamples) # averaged by codeStem
#trait.sds_byStem(stemSamples) # within codeStem variation
#trait.n_byStem(stemSamples) # number of codeStem samples that were aggregated

trait.order <- traitcol.order()
traitVars <- c(trait.order$phys.cols, trait.order$nutr.cols)
traitVars.stem <- trait.order$stem.trait.cols
traitVars.cfract <- trait.order$cfract.cols

```

Initalize lists to store results
```{r}

# save the bivariate predictor plots
bivariate_explainDecay.list <- list()

# save the fig objects that come out of makefig__...
models_explainDecay.list <- list() 

# save residuals from the best models that use traits/C fractions to explain decay
resids_explainDecay.list <- list()

# save the plots that dig into models summary results
plots_explainDecay.list <- list()

# save the model fit statistics
fitstats_explainDecay.list <- list()

```

# Slice 1: Exclude C fraction data

We expect initial wood traits will explain varitation in species+size decay rate (k and t70), species+size lagginess (alpha), and stem-level percent mass remaining at 7, 13, 25, and 37 months of decay. Specifically, we expect samples with (a) high water percent, (b) low density and total C, (c) high macro and micro nutrients, and (d) thicker bark (potential mech: limiting microbial colonization) to have faster decay and less lagginess. 

## *Hyp (species+size-level)* Species+size-level initial wood traits will predict variation decay rates and lagginess.

Compile "code-level" trait data. Check out correlations among traits and produce bivariate plots.
```{r}
traits.code %>%
  select(-c(trait.order$cfract.cols)) %>%
  filter(!is.na(C)) -> traits.code.noc
matonly <- traits.code.noc[,!colnames(traits.code.noc) %in% c(trait.order$id.cols)]
matonly.s <- scale(as.matrix(matonly))
colnames(matonly.s)

#save the bivariate plot
bivariate_explainDecay.list[['traits.code']] <- ggpairs(data = as.data.frame(matonly.s))
```

Use model selection to determine which traits to include and summarize the final models.
```{r}
result.traitsDecay <- doAnalysis_traits_explain_decayParams(decayfits, 
                                      traits.code,
                                      code.respVars, 
                                      traitVars,
                                      cfract = F) # see make_figs_woodTraits_explainDecay.R
fig <- makefig__traits_explain_decayParams(result.list = result.traitsDecay, 
                                           traitVars, cfract = F) 


fig$p
result.traitsDecay$fitstats

#save the model summary and residuals
models_explainDecay.list[['traits.code']] <- fig 
resids_explainDecay.list[['traits.code']] <- result.traitsDecay$residuals
fitstats_explainDecay.list[['traits.code']] <- result.traitsDecay$fitstats
```

- Model fits: The best fitting models used traits to explain k, t50, beta, w.t50 (all with R2 > 0.70). The worst fitting model used traits to explain ne.r2 with R2 = 0.46.

- Water content was the most influential trait on time to 50% mass loss; more water was associated with short wait times. The role of water mainly appeared to influence the shape parameter; more water led to less s-shaped decay trajectories.
```{r}
result.traitsDecay$preddat$waterperc_s %>%
  filter(respvar == c("w.t50","beta")) %>%
  mutate(respvar = factor(respvar, levels = c("w.t50","beta"), 
                         labels = c("Years to 50% mass loss",
                                    "Shape param."))) -> tmp1

p <- ggplot(tmp1, aes(x = waterperc, y = .fitted, fill = size)) +
  geom_line(aes(color = size)) +
  geom_ribbon(aes(ymin = lower, ymax = upper), alpha = .25) +
  #geom_point(aes(y = .fitted, color = size)) +
  geom_rug(sides="b", aes(color = size)) +
  labs(x = "Water content (% in fresh wood)",
       y = "Decay response") +
  theme_classic() +
  facet_wrap(~respvar, scales = "free_y")
p 


plots.water <- list(p)
```

- Ca concentration had a large influence on the shape parameter; more Ca led to more S-shaped decay trajectories. 
```{r}
result.traitsDecay$preddat$Ca_s %>%
  filter(respvar == c("w.t50","beta")) %>%
  mutate(respvar = factor(respvar, levels = c("w.t50","beta"),
                         labels = c("Years to 50% mass loss",
                                    "Shape param."))) -> tmp1

p <- ggplot(tmp1, aes(x = Ca, y = .fitted, fill = size)) +
  geom_line(aes(color = size)) +
  geom_ribbon(aes(ymin = lower, ymax = upper), alpha = .25) +
  #geom_point(aes(y = respvar, color = size)) +
  geom_rug(sides="b", aes(color = size)) +
  labs(x = "Ca (%)",
       y = "Decay response") +
  theme_classic() +
  facet_wrap(~respvar, scales = "free_y")
p

plots.ca = list(p)
```

- The largest influence on alpha was diameter size; large sized stems tended to decay at a faster rate, but see EUSC and EUTE. Also note the influence of the shape parameter in the decay model means that large stems take more time to lose 50% mass (opposite of what you'd expect)
```{r}
decayfits %>%
  select("code","species","size","alpha") -> sel.decay
sel.decay

p.size <- ggplot(sel.decay, aes(x = size, y = alpha, color = species)) +
  geom_point() +
  theme_classic() +
  geom_line(aes(group = species)) +
  ylab("Alpha") + xlab("Size")
p.size

plots.size <- list(p.size)
```

Save the plots
```{r}
plots_explainDecay.list[['traits.code']] <- list(plots.water = plots.water,
                                                 plots.ca = plots.ca,
                                                 plots.size = plots.size)

```


## *Hyp (stem-level)* Stem-level initial wood traits will predict variation in percent mass loss at each time step.

First, we need to decide what trait data (and samples) to include in this analysis since we don't have full coverage of stem-level trait data. Density and bark thickness were only measured on small sized stems.  If there is not be very much within-species variation in these traits that contribute to variation in percent mass loss than we can justify including species-level estimates of these traits in the stem-level model. 

Plot the small-sized stem-level measures of density and barkthick
```{r}
makefig__variation_densityNbarkthick(traits.stem, traits.code, pmr_byStem) # output/figures/supplementary/variation_densityNbarkthick.pdf

```

A few species have a relatively large amount of within-species variation in density (i.e. alli) and barkthickness (i.e. leer and cota).

Compare model fits (r2) using stem and species-level data to identify how much information about percent mass remaining is lost by using species-level estimates.
```{r}
cox.df <- doAnalysis_variation_densityNbarkthick(traits.stem, traits.code, pmr_byStem, stem.respVars) 
cox.df
write.csv(cox.df, file="output/tables/supplementary/smallsize_codeVstem_cox.csv")

```

Remember M1 = stem and M2 = code
- For density -- Stem-level data (M1) improves estimates of pmr after 7 and 59 months
- For barkthickness -- Stem-level data (relative to code-level data) does not improve estimates of pmr at any time point.

Compile "stem-level" trait data using (a) stem-level traits including waterperc and chemistry and (b) small species-level density and bark thickness data. Check out correlations among traits and produce bivariate plots. 
```{r}
# this creates a list of dataframes where each dataframe is for a separate model with a unique response variable
stem.respVars %>%
  purrr::set_names() %>%
  map(~CreateTraitPMRpair(respVar = .x, 
                          traits.stem, traits.code, 
                          pmr_byStem, 
                          traitVars.stem)) -> datasets # see make_figs_woodTraits_explainDecay.R

# should be able to extract just trait data from one of these dataframes
tmp <- datasets[[1]]
mat <- tmp[,colnames(tmp) %in% trait.order$stem.trait.cols]
colnames(mat)

#save the bivariate plot
bivariate_explainDecay.list[['traits.stem']] <- ggpairs(data = mat)
```

Use model selection to determine which traits to include and summarize the final models.
```{r, results=FALSE}
result.traitsPMR <- doAnalysis_traits_explain_pmr(datasets, 
                                      stem.respVars, traitVars.stem, 
                                      cfract = F) # see make_figs_woodTraits_explainDecay.R

fig <- makefig__traits_explain_pmr(result.list = result.traitsPMR, 
                                   traitVars.stem,
                                   cfract = F) 

fig$p
result.traitsPMR$fitstats

#save the model summary and residuals
models_explainDecay.list[['traits.stem']] <- fig 
resids_explainDecay.list[['traits.stem']] <- result.traitsPMR$residuals
fitstats_explainDecay.list[['traits.stem']] <- result.traitsPMR$fitstats
```

- Model fits: Models at mid time points tend to fit better, all models had R2 ~ 50%

Set up added variable plot dataframes

- Water content was the most influential trait, especially for explaining mass loss at mid to late decay stages (25-59 months); More water led to lower percent mass remaining
- Size was also influential, especially for explaining mass loss at earlier decay stages (7 and 13 months); large stems had greater percent mass remaining during early time points
```{r}
result.traitsPMR$preddat$waterperc_s %>%
  mutate(respvar = factor(respvar, 
                          levels = c("time7","time13","time25","time37","time59"))) -> tmp1
p <- ggplot(tmp1, aes(x = waterperc, y = .fitted, fill = size)) +
  geom_line(aes(color = size)) +
  geom_ribbon(aes(ymin = lower, ymax = upper), alpha = .25) +
  #geom_point(aes(y = respvar, color = size)) +
  geom_rug(sides="b", aes(color = size)) +
  labs(x = "Water content (% in fresh wood)",
       y = "% Mass remaining after X months") +
  theme_classic() +
  facet_grid(~respvar)
p

plots <- list(p)
```

Save the plots
```{r}
plots_explainDecay.list[['traits.stem']] <- list(plots = plots)
```


                                                 

# Slice 2: Include C fraction data (and exclude small stems)

## *Hyp (species+size-level)* Species+size-level initial C fractions will predict variation decay rates and lagginess.

Compile "code-level" C fraction trait data. Check out correlations among traits and produce bivariate plots. 
```{r}
traits.code %>%
  select(c(trait.order$id.cols, trait.order$cfract.cols, "waterperc","N","Ca")) %>%
  filter(!is.na(Arabinose)) -> traits.code.cfract
sum(!complete.cases(traits.code.cfract)) # no missing trait data

#save the bivariate plot
bivariate_explainDecay.list[['cfract.code']] <- ggpairs(data =
                                                      traits.code.cfract[,trait.order$cfract.cols])
```

Use model selection to determine which Cfracts to include and summarize the final models.
```{r}
result.cfractDecay <- doAnalysis_traits_explain_decayParams(decayfits, 
                                      traits.code = traits.code.cfract,
                                      code.respVars, 
                                      traitVars = traitVars.cfract,
                                      cfract = T) # see make_figs_woodTraits_explainDecay.R

fig <- makefig__traits_explain_decayParams(result.list = result.cfractDecay, 
                                           traitVars = traitVars.cfract, 
                                           cfract = T) 


fig$p
result.cfractDecay$fitstats

#save the model summary and residuals
models_explainDecay.list[['cfract.code']]<- fig
resids_explainDecay.list[['cfract.code']] <- result.cfractDecay$residuals
fitstats_explainDecay.list[['cfract.code']] <- result.cfractDecay$fitstats
```

- Model fits: Over 50% of variance in decay R2 and time to 50% mass loss was explained by C fractions. Less more explanatory power for beta than alpha.
- Lignin was the most influential trait, especially for model fit; less lignin led to better fitting models. More lignin led to longer wait times, slower decay rates (alpha), and more S-shaped decay (beta).
```{r}
result.cfractDecay$preddat$Lignin_s %>%
  mutate(respvar = factor(respvar, levels = c("w.t50","w.r2","alpha","beta"),
                          labels = c("Years to 50% mass loss",
                                     "Weibull R2",
                                     "Scale param.",
                                     "Shape param."))) -> tmp1

p <- ggplot(tmp1, aes(x = Lignin, y = .fitted, fill = size)) +
  geom_line(aes(color = size)) +
  geom_ribbon(aes(ymin = lower, ymax = upper), alpha = .25) +
  #geom_point(aes(y = .fitted, color = size)) +
  geom_rug(sides="b", aes(color = size)) +
  labs(x = "Lignin (%)",
       y = "Decay response") +
  theme_classic() +
  facet_wrap(~respvar, scales = "free_y")
p
plots.l <- list(p)

```

- Rhamnose and Glucose were important in explaining variance in wait times and the shape parameter beta. More rhamanose and glucose led to longer wait times and more s-shaped decay trajectories.
```{r}
result.cfractDecay$preddat$Rhamnose_s %>%
  filter(respvar %in% c("w.t50","beta")) %>%
  mutate(respvar = factor(respvar, levels = c("w.t50","beta"),
                          labels = c("Years to 50% mass loss",
                                     "Shape param."))) -> tmp1

result.cfractDecay$preddat$Glucose_s %>%
  filter(respvar %in% c("w.t50","beta")) %>%
  mutate(respvar = factor(respvar, levels = c("w.t50","beta"),
                          labels = c("Years to 50% mass loss",
                                     "Shape param."))) -> tmp2

p.1 <- ggplot(tmp1, aes(x = Rhamnose, y = .fitted, fill = size)) +
  geom_line(aes(color = size)) +
  geom_ribbon(aes(ymin = lower, ymax = upper), alpha = .25) +
  #geom_point(aes(y = .fitted, color = size)) +
  geom_rug(sides="b", aes(color = size)) +
  labs(x = "Rhamnose (%)",
       y = "Decay response") +
  theme_classic() +
  facet_wrap(~respvar, scales = "free_y")
p.1

p.2 <- ggplot(tmp2, aes(x = Glucose, y = .fitted, fill = size)) +
  geom_line(aes(color = size)) +
  geom_ribbon(aes(ymin = lower, ymax = upper), alpha = .25) +
  #geom_point(aes(y = .fitted, color = size)) +
  geom_rug(sides="b", aes(color = size)) +
  labs(x = "Glucose (%)",
       y = "Decay response") +
  theme_classic() +
  facet_wrap(~respvar, scales = "free_y")
p.2

grid.arrange(p.1, p.2, ncol = 1)
plots.rhgl <- list(p.1, p.2)

```

Save the plots
```{r}
plots_explainDecay.list[['cfract.code']] <- list(plots.l = plots.l,
                                                 plots.rhgl = plots.rhgl)
```


## *Hyp (stem-level)* Stem-level initial C fractions will predict variation in percent mass loss at each time step.

Compile "stem-level" C fraction trait data. Check out correlations among traits and produce bivariate plots. 
```{r}
traits.stem %>%
  select(c(trait.order$id.cols, "codeStem", trait.order$cfract.cols)) %>%
  filter(!is.na(Arabinose)) -> traits.stem.cfract
sum(!complete.cases(traits.stem.cfract)) # no missing trait data

#save the bivariate plot
bivariate_explainDecay.list[['cfract.stem']] <- ggpairs(data =
                                                      traits.stem.cfract[,trait.order$cfract.cols])
```

Use model selection to determine which Cfracts to include and summarize the final models.
```{r}
# this creates a list of dataframes where each dataframe is for a separate model with a unique response variable
stem.respVars %>%
  purrr::set_names() %>%
  map(~CreateCfractPMRpair(respVar = .x, 
                          traits.stem.cfract, 
                          pmr_byStem, 
                          traitVars.cfract)) -> datasets # see make_figs_woodTraits_explainDecay.R

# do analysis
result.cfractPMR <- doAnalysis_traits_explain_pmr(datasets, 
                                      stem.respVars, 
                                      traitVars.stem = traitVars.cfract,
                                      cfract = T) # see make_figs_woodTraits_explainDecay.R

fig <- makefig__traits_explain_pmr(result.list = result.cfractPMR, 
                                   traitVars.stem = traitVars.cfract,
                                   cfract = T) 

fig$p
result.cfractPMR$fitstats

#save the model summary and residuals
models_explainDecay.list[['cfract.stem']] <- fig 
resids_explainDecay.list[['cfract.stem']] <- result.cfractPMR$residuals
fitstats_explainDecay.list[['cfract.stem']] <- result.cfractPMR$fitstats
```

- Model fits: Relatively poor model fits. Maximum R2 is 0.45 and that is for time25. Minimum R2 is .17 which is for time7

- Arabinose is the only influential fraction for time7; more arabinose led to less mass remaining. Looking at the plot, this is could be driven by the high arabinose low mass remaining point...
```{r}
result.cfractPMR$preddat$Arabinose_s %>%
  mutate(respvar = factor(respvar, levels = c("time7","time13","time25","time37","time59"))) -> tmp

p.early <- ggplot(tmp, aes(x = Arabinose, y = .fitted, fill = size)) +
  geom_line(aes(color = size)) +
  geom_ribbon(aes(ymin = lower, ymax = upper), alpha = .25) +
  #geom_point(aes(y = .fitted, color = size)) +
  geom_rug(sides="b", aes(color = size)) +
  labs(x = "Arabinose (%)",
       y = "% Mass remaining after X months") +
  theme_classic() +
  facet_grid(~respvar)
p.early

plots.early <- list(p.early)
```

- Xylose is the most influential fraction for time25; more xylose led to less mass remaining. Rhamnose was also influential at timethis time point; more rhamnose led to more mass remaining. Mannose was the most influential fraction for time 37; more mannose led to more mass remaining.
```{r}
result.cfractPMR$preddat$Xylose_s %>%
  mutate(respvar = factor(respvar, levels = c("time7","time13","time25","time37","time59"))) %>%
  filter(respvar == "time25")-> tmp

p.1 <- ggplot(tmp, aes(x = Xylose, y = .fitted, fill = size)) +
  geom_line(aes(color = size)) +
  geom_ribbon(aes(ymin = lower, ymax = upper), alpha = .25) +
  #geom_point(aes(y = respval, color = size)) +
  geom_rug(sides="b", aes(color = size)) +
  labs(x = "Xylose (%)",
       y = "% Mass remaining after X months") +
  theme_classic() +
  facet_grid(~respvar)
p.1

result.cfractPMR$preddat$Rhamnose_s %>%
  mutate(respvar = factor(respvar, levels = c("time7","time13","time25","time37","time59"))) %>%
  filter(respvar %in% c("time25"))-> tmp

p.2 <- ggplot(tmp, aes(x = Rhamnose, y = .fitted, fill = size)) +
  geom_line(aes(color = size)) +
  geom_ribbon(aes(ymin = lower, ymax = upper), alpha = .25) +
  #geom_point(aes(y = .fitted, color = size)) +
  geom_rug(sides="b", aes(color = size)) +
  labs(x = "Rhamnose (%)",
       y = "% Mass remaining after X months") +
  theme_classic() +
  facet_grid(~respvar)
p.2

result.cfractPMR$preddat$Rhamnose_s %>%
  mutate(respvar = factor(respvar, levels = c("time7","time13","time25","time37","time59"))) %>%
  filter(respvar %in% c("time37"))-> tmp
p.3 <- ggplot(tmp, aes(x = Rhamnose, y = .fitted, fill = size)) +
  geom_line(aes(color = size)) +
  geom_ribbon(aes(ymin = lower, ymax = upper), alpha = .25) +
  #geom_point(aes(y = .fitted, color = size)) +
  geom_rug(sides="b", aes(color = size)) +
  labs(x = "Rhamnose (%)",
       y = "% Mass remaining after X months") +
  theme_classic() +
  facet_grid(~respvar)
p.3

result.cfractPMR$preddat$Mannose_s %>%
  mutate(respvar = factor(respvar, levels = c("time7","time13","time25","time37","time59"))) %>%
  filter(respvar == "time37")-> tmp

p.4 <- ggplot(tmp, aes(x = Mannose, y = .fitted, fill = size)) +
  geom_line(aes(color = size)) +
  geom_ribbon(aes(ymin = lower, ymax = upper), alpha = .25) +
  #geom_point(aes(y = .fitted, color = size)) +
  geom_rug(sides="b", aes(color = size)) +
  labs(x = "Mannose (%)",
       y = "% Mass remaining after X months") +
  theme_classic() +
  facet_grid(~respvar)
p.4

grid.arrange(p.1, p.2, p.3, p.4)
plots.mid <- list(p.1, p.2, p.3, p.4)

```

- Lignin was the most influential fraction for time59 (and was generally more informative at the last two time points); more lignin led to more mass remaining.
```{r}

result.cfractPMR$preddat$Lignin_s %>%
  mutate(respvar = factor(respvar, levels = c("time7","time13","time25","time37","time59"))) -> tmp

p <- ggplot(tmp, aes(x = Lignin, y = .fitted, fill = size)) +
  geom_line(aes(color = size)) +
  geom_ribbon(aes(ymin = lower, ymax = upper), alpha = .25) +
  #geom_point(aes(y = respval, color = size)) +
  geom_rug(sides="b", aes(color = size)) +
  labs(x = "Lignin (%)",
       y = "% Mass remaining after X months") +
  theme_classic() +
  facet_grid(~respvar)
p

plots.l <- list(p)

```

Save the plots
```{r}
plots_explainDecay.list[['cfract.stem']] <- list(plots.early = plots.early,
                                                 plots.mid = plots.mid,
                                                 plots.l = plots.l)
```


# Print results

Save bivariate predictor plots
```{r}
# save the bivariate predictor plots
names(bivariate_explainDecay.list)
#uncomment if data change
# ggsave(filename = "output/figures/supplementary/bivariate_traits-code.pdf",
#        plot = bivariate_explainDecay.list$traits.code, width = 10, height = 10)
# ggsave(filename = "output/figures/supplementary/bivariate_traits-stem.pdf",
#        plot = bivariate_explainDecay.list$traits.stem, width = 10, height = 10)
# ggsave(filename = "output/figures/supplementary/bivariate_cfract-code.pdf",
#        plot = bivariate_explainDecay.list$cfract.code, width = 10, height = 10)
# ggsave(filename = "output/figures/supplementary/bivariate_cfract-stem.pdf",
#        plot = bivariate_explainDecay.list$cfract.code, width = 10, height = 10)
```

Save best model summaries
```{r}
# save the objects that come out of makefig__...
getrid.of.stuff <- function(plot.obj){
  p <- plot.obj +
  scale_fill_distiller(direction = 1, limits = c(0, .6)) +
  theme(axis.title=element_blank(),
        axis.text=element_blank(),
        axis.ticks=element_blank()) +
  guides(fill = F)
  
  return(p)
}

p.tc <- getrid.of.stuff(models_explainDecay.list$traits.code$p)
p.ts <- getrid.of.stuff(models_explainDecay.list$traits.stem$p)
p.cc <- getrid.of.stuff(models_explainDecay.list$cfract.code$p)
p.cs <- getrid.of.stuff(models_explainDecay.list$cfract.stem$p)

# extract axes
y.t <- models_explainDecay.list$traits.code$p +
  theme(axis.title.x=element_blank(),
        axis.text.x=element_blank(),
        axis.ticks.x=element_blank(),
        axis.title.y=element_blank()) + guides(fill = F)
y.c <- models_explainDecay.list$cfract.code$p +
  theme(axis.title.x=element_blank(),
        axis.text.x=element_blank(),
        axis.ticks.x=element_blank(),
        axis.title.y=element_blank()) + guides(fill = F)
x.c <- models_explainDecay.list$traits.code$p +
  theme(axis.title.x=element_blank(),
        axis.text.y=element_blank(),
        axis.ticks.y=element_blank(),
        axis.title.y=element_blank()) + guides(fill = F)
x.s <- models_explainDecay.list$cfract.code$p +
  theme(axis.title.x=element_blank(),
        axis.text.y=element_blank(),
        axis.ticks.y=element_blank(),
        axis.title.y=element_blank()) + guides(fill = F)
require(gtable)
p.y1 <- gtable_filter(ggplotGrob(y.t), 'axis-l', trim=F)
p.y2 <- gtable_filter(ggplotGrob(y.c), 'axis-l', trim=F)
p.x1 <- gtable_filter(ggplotGrob(x.c), 'axis-b', trim=F)
p.x2 <- gtable_filter(ggplotGrob(x.s), 'axis-b', trim=F)
empty <- textGrob("")

pdf("output/figures/maintext/traits-cfract_explainDecay.pdf", width = 6, height = 6)
grid.arrange(p.y1, p.tc, p.ts, 
             p.y2, p.cc, p.cs, 
             empty, p.x1, p.x2, ncol = 3,
             widths = c(1.75, 4, 5),
             heights = c(12, 7, 3))
dev.off()

# print the fill legend separately
new.p <- models_explainDecay.list$traits.code$p +
           scale_fill_distiller(direction = 1, limits = c(0, .6))
tmp <- gtable_filter(ggplotGrob(new.p), 'guide-box', trim=F)
pdf("output/figures/maintext/traits-cfract_explainDecay_FILLGUIDE.pdf", width = 6, height = 6)
grid.arrange(tmp)
dev.off()

```

Save plots to explore model results
```{r}
# save the plots that dig into models summary results
names(plots_explainDecay.list)
```

traits.code
```{r}
#traits.code
names(plots_explainDecay.list$traits.code)

## water
pdf(file = "output/unknown/traits-code_explainDecay_waterCa.pdf")
grid.arrange(plots_explainDecay.list$traits.code$plots.water[[1]] + guides(color = F, fill = F),
             plots_explainDecay.list$traits.code$plots.ca[[1]] + guides(color = F, fill = F))
dev.off()

## size
pdf(file = "output/unknown/traits-code_explainDecay_size.pdf")
plots_explainDecay.list$traits.code$plots.size[[1]]
dev.off()

```

traits.stem
```{r}
names(plots_explainDecay.list$traits.stem)

pdf(file = "output/unknown/traits-stem_explainDecay_water.pdf")
plots_explainDecay.list$traits.stem$plots[[1]]
dev.off()
```

cfract.code
```{r}
names(plots_explainDecay.list$cfract.code)

## lignin
pdf(file = "output/unknown/cfract-code_explainDecay_lignin.pdf")
plots_explainDecay.list$cfract.code$plots.l[[1]]
dev.off()

## rhgl
pdf(file = "output/unknown/cfract-code_explainDecay_rhgl.pdf")
grid.arrange(plots_explainDecay.list$cfract.code$plots.rhgl[[1]],
             plots_explainDecay.list$cfract.code$plots.rhgl[[2]], ncol = 1)
dev.off()

```

cfract.stem
```{r}
names(plots_explainDecay.list$cfract.stem)

## early
pdf(file = "output/unknown/cfract-stem_explainDecay_early.pdf")
plots_explainDecay.list$cfract.stem$plots.early[[1]]
dev.off()

## mid
pdf(file = "output/unknown/cfract-stem_explainDecay_mid.pdf")
grid.arrange(plots_explainDecay.list$cfract.stem$plots.mid[[1]] + guides(color = F, fill = F),
             plots_explainDecay.list$cfract.stem$plots.mid[[4]] + guides(color = F, fill = F),
             plots_explainDecay.list$cfract.stem$plots.mid[[2]] + guides(color = F, fill = F),
             plots_explainDecay.list$cfract.stem$plots.mid[[3]] + guides(color = F, fill = F),ncol = 2)
dev.off()

## lignin
pdf(file = "output/unknown/cfract-stem_explainDecay_lignin.pdf")
plots_explainDecay.list$cfract.stem$plots.l[[1]]
dev.off()

```

Save model residuals
```{r}
# save residuals from the best models that use traits/C fractions to explain decay
names(resids_explainDecay.list)
saveRDS(resids_explainDecay.list, file = "derived_data/resids_explainDecay_list.RData")

```

Save model fit statistics
```{r}
fnames <- paste0("output/tables/", names(fitstats_explainDecay.list),".csv")
map2(.x = fitstats_explainDecay.list, .y = fnames,
     ~write.csv(.x, file = .y))
```
