==> 3c_endoDiv_explainDecay.Rmd <==
---
title: "Does endophyte diversity explain decay?"
author: "Marissa Lee"
date: "12/2/2018"
output: github_document
---

Contents:
1. CODE-S1. Test role of endophyte diversity. Exclude C fraction data, include small stems.
2. CODE-S2. Test role of endophyte diversity. C fraction data
3. STEM-S1. Test role of endophyte diversity. Exclude C fraction data, include small stems.
4. STEM-S2. Test role of endophyte diversity. C fraction data

1-r. CODE-S1. Test role of endophyte diversity after accounting for traits. Exclude C fraction data, include small stems.
2-r. CODE-S2. Test role of endophyte diversity after accounting for traits. C fraction data.
3-r. STEM-S1. Test role of endophyte diversity after accounting for traits. Exclude C fraction data, include small stems.
4-r. STEM-S2. Test role of endophyte diversity after accounting for traits. C fraction data.

**Note that the full community matrix was used for all analyses**

___________________
___________________

Load libraries and functions
```{r}
#chunk options
knitr::opts_chunk$set(echo=FALSE, message=FALSE, warning=FALSE)

#---------------------------#
#libraries
library(tidyverse) # ggplot2, dplyr, tidyr, readr, purrr
library(easystats) # report, performance, parameters, correlation, effectsize
library(rioja) # for WAPLS

# commented out because I'm going to try using easystats instead...
#library(tidymodels) # broom
#library(sjstats) #provides function eta_sq() to extract r2 for each predictor in a lm
#library(lmtest) # for coxtest()

# commented out because I'm going to try using cowplot instead
#library(grid) #plotting
#library(gridExtra) # plotting
#library(GGally) # bivariate plots


#---------------------------#
#fxns
source('code/helper_fxns.R')
sourceDir('code')

```

Load decay data
```{r loaddata}

# samples and decay
stemSamples <- load_stemSamples() # this function is in load_microbeData.R
initial_mass <- read_in_initial_mass()
harvest_mass <- LoadHarvestFiles()
pmr <- Calc_massRemaining(initial_mass, harvest_mass) # calculate percent mass remaining (pmr)
pmr_byStem <- AvgPMR_byStem(pmr, long.form = F) # average pmr for each stem and timepoint
stem.respVars <- list("time7", "time13", "time25", "time37","time59")
decayfits <- read_csv("derived_data/decayfits.csv") # see 1_decayPatterns.Rmd to update
code.respVars <- list("w.t50","beta","alpha","w.r2")
# weibull params
# beta = scale param = higher values mean *slower* decay
# alpha = shape param = higher values mean more S-shaped
```

Load wood trait data
```{r}
traits.code <- read.csv(file = "derived_data/traits_code.csv", row.names = 1)
traits.stem <- read.csv(file = "derived_data/traits_stem.csv", row.names = 1)

# ordering and stuff
trait.order <- traitcol.order()
traitVars1 <- c(trait.order$phys.cols, trait.order$nutr.cols)
traitVars1 # slice1 (w/ size)
traitVars2 <- c(trait.order$cfract.cols)
traitVars2 # slice2 (w/o size)

```

Load residuals from 2b_woodTraits_explainDecay.Rmd
```{r}
resids_explainDecay.list <- readRDS(file = "derived_data/resids_explainDecay_list.RData")

```

Load OTU tables
```{r}
otu.list <- readRDS(file = "derived_data/otu_list.RData")
otu.list.trimmed <- readRDS(file = "derived_data/otu_list_trimmed.RData")

# microbes
microb.data <- load_MicrobeCollection(stemSamples)
taxAndFunguild <- microb.data$taxAndFunguild
comm.otu <- microb.data$comm.otu
seqSamples <- microb.data$seqSamples

```


___________________
# 1. CODE-S1. Test role of endophyte diversity. Exclude C fraction data, include small stems.

calculate diversity metrics and set up dataframe
```{r}
calc.div <- function(comm.otu){
  require(vegan)
  H <- diversity(comm.otu) # shannon's h
  comm.pa <- (comm.otu>0)*1
  SR <- rowSums(comm.pa) # species richness
  df <- data.frame(seq_sampName = names(SR), SR, H, 
                   stringsAsFactors = F, row.names = NULL)
  
  return(df)
}

otu.list %>%
  map(~calc.div(.x)) -> div.list

# richness only
div.list %>%
  map(~left_join(.x, seqSamples)) %>%
  list_to_df() %>%
  dplyr::rename('community'='source') %>%
  group_by(code, community) %>%
  summarize(mean=mean(SR),
            se=sd(SR)/sqrt(length(SR)),
            n = sum(!is.na(SR)),
            upper=mean+se,
            lower=mean-se) -> div.df
div.df %>%
  left_join(decayfits) -> rich.df
write.csv(rich.df, file = "derived_data/richness_C.csv")
```

fit models
```{r}
unique(rich.df$community)
fit.mods.c1 <- function(rich.df, curr.comm){
  
  rich.df %>%
    filter(community == curr.comm) -> df
  mod1 <- lm(w.t50 ~ size * mean, data = df)
  mod2 <- lm(beta ~ size * mean, data = df)
  mod3 <- lm(alpha ~ size * mean, data = df)
  mod4 <- lm(w.r2 ~ size * mean, data = df)
  mod.list <- list(mod1, mod2, mod3, mod4)
  names(mod.list) <- unlist(code.respVars)
  # model r2s
  mod.list %>%
    map(~r2(.x)) %>%
    map_df(.f = unlist, use.names = T, .id = "respvar") %>%
    set_names(nm = c("respvar","Rsq","Rsq.adj")) %>%
    mutate(community = curr.comm) -> df.r2
  df.r2
  # anova table
  mod.list %>%
    map(~aov(.x)) %>%
    map(~model_parameters(.x, eta_squared = "partial")) %>%
    map_df(.f = tibble, .id = "respvar") -> df.an
  # coefs
  mod.list %>%
    map(~model_parameters(.x)) %>%
    map_df(.f = tibble, .id = "respvar") -> df.coefs # coefs
  
  result.list <- list(r2 = df.r2, an = df.an, coefs = df.coefs)
  return(result.list)
  
}

unique(rich.df$community) %>%
  set_names() %>%
  map(~fit.mods.c1(rich.df = rich.df, curr.comm = .x)) -> out.c1
out.c1
out.c1 %>%
  map(~tibble(.x)) %>%
  map(~unnest(.x, cols= c(.x))) %>%
  map_df(.f = tibble, .id = "comm") %>%
  arrange(comm, respvar) -> out.c1.df
out.c1.df

# summarize signif models
out.c1.df %>%
  filter(Parameter %in% c("mean","size:mean")) %>%
  filter(p < 0.05) -> signif.c1.df
signif.c1.df

out.c1.df %>%
  filter(respvar == "alpha") %>%
  filter(comm %in% c("all","basid","patho","sapro")) %>%
  filter(!is.na(Coefficient)) %>%
  select(comm, respvar, Parameter, p, Coefficient, SE) -> signifcoef.c1.df
signifcoef.c1.df

```

___________________
# 2. CODE-S2. Test role of endophyte diversity. C fraction data

fit models
```{r}
rich.df %>%
  filter(size == "large") -> rich.df2
unique(rich.df2$community)

fit.mods.c2 <- function(rich.df2, curr.comm){
  
  rich.df2 %>%
    filter(community == curr.comm) -> df
  mod1 <- lm(w.t50 ~ mean, data = df)
  mod2 <- lm(beta ~ mean, data = df)
  mod3 <- lm(alpha ~ mean, data = df)
  mod4 <- lm(w.r2 ~ mean, data = df)
  mod.list <- list(mod1, mod2, mod3, mod4)
  names(mod.list) <- unlist(code.respVars)
  
  # model r2s
  mod.list %>%
    map(~r2(.x)) %>%
    map_df(.f = unlist, use.names = T, .id = "respvar") %>%
    set_names(nm = c("respvar","Rsq","Rsq.adj")) %>%
    mutate(community = curr.comm) -> df.r2
  df.r2
  # anova table
  mod.list %>%
    map(~aov(.x)) %>%
    map(~model_parameters(.x, eta_squared = "partial")) %>%
    map_df(.f = tibble, .id = "respvar") -> df.an
  # coefs
  mod.list %>%
    map(~model_parameters(.x)) %>%
    map_df(.f = tibble, .id = "respvar") -> df.coefs # coefs

  result.list <- list(r2 = df.r2, an = df.an, coefs = df.coefs)
  return(result.list)
  
}


unique(rich.df2$community) %>%
  set_names() %>%
  map(~fit.mods.c2(rich.df2 = rich.df2, curr.comm = .x)) -> out.c2
out.c2 %>%
  map(~tibble(.x)) %>%
  map(~unnest(.x, cols= c(.x))) %>%
  map_df(.f = tibble, .id = "comm") %>%
  arrange(comm, respvar) -> out.c2.df
out.c2.df

# summarize signif models
out.c2.df %>%
  filter(Parameter %in% c("mean")) %>%
  filter(p < 0.05) -> signif.c2.df
signif.c2.df
out.c2.df %>%
  filter(p<0.05)

```


___________________
# 3. STEM-S1. Test role of endophyte diversity. Exclude C fraction data, include small stems.

calculate diversity metrics and set up dataframe
```{r}
calc.div <- function(comm.otu){
  require(vegan)
  H <- diversity(comm.otu) # shannon's h
  comm.pa <- (comm.otu>0)*1
  SR <- rowSums(comm.pa) # species richness
  df <- data.frame(seq_sampName = names(SR), SR, H, 
                   stringsAsFactors = F, row.names = NULL)
  
  return(df)
}

otu.list %>%
  map(~calc.div(.x)) -> div.list

# richness only
div.list %>%
  map(~left_join(.x, seqSamples)) %>%
  list_to_df() %>%
  dplyr::rename('community'='source') %>%
  group_by(community, codeStem) %>%
  summarize(mean=mean(SR),
            se=sd(SR)/sqrt(length(SR)),
            n = sum(!is.na(SR)),
            upper=mean+se,
            lower=mean-se) -> div.df
head(div.df)
div.df %>%
  left_join(pmr_byStem) -> rich.df
head(rich.df)
write.csv(rich.df, file = "derived_data/richness_S.csv")
```

fit models
```{r}
unique(rich.df$community)
stem.respVars

fit.mods.s1 <- function(rich.df, curr.comm){
  
  rich.df %>%
    filter(community == curr.comm) -> df
  mod1 <- lm(time7 ~ size * mean, data = df)
  mod2 <- lm(time13 ~ size * mean, data = df)
  mod3 <- lm(time25 ~ size * mean, data = df)
  mod4 <- lm(time37 ~ size * mean, data = df)
  mod5 <- lm(time59 ~ size * mean, data = df)
  mod.list <- list(mod1, mod2, mod3, mod4, mod5)
  names(mod.list) <- unlist(stem.respVars)
  
  # model r2s
  mod.list %>%
    map(~r2(.x)) %>%
    map_df(.f = unlist, use.names = T, .id = "respvar") %>%
    set_names(nm = c("respvar","Rsq","Rsq.adj")) %>%
    mutate(community = curr.comm) -> df.r2
  df.r2
  # anova table
  mod.list %>%
    map(~aov(.x)) %>%
    map(~model_parameters(.x, eta_squared = "partial")) %>%
    map_df(.f = tibble, .id = "respvar") -> df.an
  # coefs
  mod.list %>%
    map(~model_parameters(.x)) %>%
    map_df(.f = tibble, .id = "respvar") -> df.coefs # coefs
  
  result.list <- list(r2 = df.r2, an = df.an, coefs = df.coefs)
  return(result.list)
  
}

unique(rich.df$community) %>%
  set_names() %>%
  map(~fit.mods.s1(rich.df = rich.df, curr.comm = .x)) -> out.s1
out.s1
out.s1 %>%
  map(~tibble(.x)) %>%
  map(~unnest(.x, cols= c(.x))) %>%
  map_df(.f = tibble, .id = "comm") %>%
  arrange(comm, respvar) -> out.s1.df

out.s1.df
# summarize signif models
out.s1.df %>%
  filter(Parameter %in% c("mean","size:mean")) %>%
  filter(p < 0.05) -> signif.s1.df
signif.s1.df

out.s1.df %>%
  filter(comm == "patho" & respvar == "time37") %>%
  filter(!is.na(p)) %>%
  select(comm, respvar, Parameter, Coefficient, p) %>%
  mutate(p = round(p, digits = 3))

```

examine interactions
```{r}
rich.df %>%
    filter(community == "patho") -> df
mod <- lm(time37 ~ size * mean, data = df)
model_parameters(mod) 
estimate_slopes(mod)


```

___________________
# 4. STEM-S2. Test role of endophyte diversity. C fraction data

fit models
```{r}
rich.df %>%
  filter(size == "large") -> rich.df2

fit.mods.s2 <- function(rich.df2, curr.comm){
  
  rich.df2 %>%
    filter(community == curr.comm) -> df
  mod1 <- lm(time7 ~ mean, data = df)
  mod2 <- lm(time13 ~ mean, data = df)
  mod3 <- lm(time25 ~ mean, data = df)
  mod4 <- lm(time37 ~ mean, data = df)
  mod5 <- lm(time59 ~ mean, data = df)
  mod.list <- list(mod1, mod2, mod3, mod4, mod5)
  names(mod.list) <- unlist(stem.respVars)
  
  # model r2s
  mod.list %>%
    map(~r2(.x)) %>%
    map_df(.f = unlist, use.names = T, .id = "respvar") %>%
    set_names(nm = c("respvar","Rsq","Rsq.adj")) %>%
    mutate(community = curr.comm) -> df.r2
  df.r2
  # anova table
  mod.list %>%
    map(~aov(.x)) %>%
    map(~model_parameters(.x, eta_squared = "partial")) %>%
    map_df(.f = tibble, .id = "respvar") -> df.an
  # coefs
  mod.list %>%
    map(~model_parameters(.x)) %>%
    map_df(.f = tibble, .id = "respvar") -> df.coefs # coefs
  
  result.list <- list(r2 = df.r2, an = df.an, coefs = df.coefs)
  return(result.list)
  
}

unique(rich.df2$community) %>%
  set_names() %>%
  map(~fit.mods.s2(rich.df2 = rich.df2, curr.comm = .x)) -> out.s2
out.s2
out.s2 %>%
  map(~tibble(.x)) %>%
  map(~unnest(.x, cols= c(.x))) %>%
  map_df(.f = tibble, .id = "comm") %>%
  arrange(comm, respvar) -> out.s2.df

# summarize signif models
out.s2.df %>%
  filter(Parameter %in% c("mean","size:mean")) %>%
  filter(p < 0.05) -> signif.s2.df
signif.s2.df

out.s2.df %>%
  filter(comm == "oo" & respvar == "time25") %>%
  filter(!is.na(p)) %>%
  select(comm, respvar, Parameter, Coefficient, p) %>%
  mutate(p = round(p, digits = 3))

```


##############################################

___________________
# 1-r. CODE-S1. Test role of endophyte diversity after accounting for traits. Exclude C fraction data, include small stems.

calculate diversity metrics and set up dataframe
```{r}
otu.list %>%
  map(~calc.div(.x)) -> div.list

# richness only
div.list %>%
  map(~left_join(.x, seqSamples)) %>%
  list_to_df() %>%
  dplyr::rename('community'='source') %>%
  group_by(code, community) %>%
  summarize(mean=mean(SR),
            se=sd(SR)/sqrt(length(SR)),
            n = sum(!is.na(SR)),
            upper=mean+se,
            lower=mean-se) -> div.df
head(div.df)

resids_explainDecay.list <- readRDS(file = "derived_data/resids_explainDecay_list.RData")
residfits.c1<- resids_explainDecay.list[['C1']]
residfits.c1 %>%
  spread(key = resp, value = resid) -> residfits.c1.w
head(residfits.c1.w)

size.indx <- unique(seqSamples[,c("code","size")])
div.df %>%
  left_join(residfits.c1.w) %>%
  left_join(size.indx)-> rich.df
head(rich.df)
write.csv(rich.df, file = "derived_data/richness_C1_resid.csv")
```

fit models
```{r}
rich.df <- read.csv("derived_data/richness_C1_resid.csv", row.names = 1)
head(rich.df)
unique(rich.df$community)

fit.mods.c1 <- function(rich.df, curr.comm){
  
  rich.df %>%
    filter(community == curr.comm) -> df
  mod1 <- lm(w.t50 ~ size * mean, data = df)
  mod2 <- lm(beta ~ size * mean, data = df)
  mod3 <- lm(alpha ~ size * mean, data = df)
  mod4 <- lm(w.r2 ~ size * mean, data = df)
  mod.list <- list(mod1, mod2, mod3, mod4)
  names(mod.list) <- unlist(code.respVars)
  # model r2s
  mod.list %>%
    map(~r2(.x)) %>%
    map_df(.f = unlist, use.names = T, .id = "respvar") %>%
    set_names(nm = c("respvar","Rsq","Rsq.adj")) %>%
    mutate(community = curr.comm) -> df.r2
  df.r2
  # anova table
  mod.list %>%
    map(~aov(.x)) %>%
    map(~model_parameters(.x, eta_squared = "partial")) %>%
    map_df(.f = tibble, .id = "respvar") -> df.an
  # coefs
  mod.list %>%
    map(~model_parameters(.x)) %>%
    map_df(.f = tibble, .id = "respvar") -> df.coefs # coefs
  
  result.list <- list(r2 = df.r2, an = df.an, coefs = df.coefs)
  return(result.list)
  
}

unique(rich.df$community) %>%
  set_names() %>%
  map(~fit.mods.c1(rich.df = rich.df, curr.comm = .x)) -> out.c1
out.c1 %>%
  map(~tibble(.x)) %>%
  map(~unnest(.x, cols= c(.x))) %>%
  map_df(.f = tibble, .id = "comm") %>%
  arrange(comm, respvar) -> out.c1.df

# summarize signif models
out.c1.df %>%
  filter(Parameter %in% c("mean","size:mean")) %>%
  filter(p < 0.05) -> signif.c1.df
signif.c1.df
out.c1.df %>%
  filter(respvar == "alpha") %>%
  filter(comm %in% c("all")) %>%
  select(comm, respvar, Parameter, p, Coefficient, SE)



```

examine interactions
```{r}
rich.df %>%
    filter(community == "sapro") -> df
mod <- lm(w.t50 ~ size * mean, data = df)
model_parameters(mod) 
estimate_slopes(mod)

head(df)
new.data <- data.frame(size = c("large","small"), 
                       mean = mean(df$mean))
new.data
df$pred <- predict(mod)

ggplot(df, aes(x = mean, y = w.t50, color = size)) +
  geom_point() +
  geom_line(aes(y = pred))



```
___________________
# 2-r. CODE-S2. Test role of endophyte diversity after accounting for traits. C fraction data.

calculate diversity metrics and set up dataframe
```{r}
otu.list %>%
  map(~calc.div(.x)) -> div.list

# richness only
div.list %>%
  map(~left_join(.x, seqSamples)) %>%
  list_to_df() %>%
  dplyr::rename('community'='source') %>%
  group_by(code, community) %>%
  summarize(mean=mean(SR),
            se=sd(SR)/sqrt(length(SR)),
            n = sum(!is.na(SR)),
            upper=mean+se,
            lower=mean-se) -> div.df

resids_explainDecay.list <- readRDS(file = "derived_data/resids_explainDecay_list.RData")
residfits.c2<- resids_explainDecay.list[['C2']]
residfits.c2 %>%
  spread(key = resp, value = resid) -> residfits.c2.w

size.indx <- unique(seqSamples[,c("code","size")])
div.df %>%
  left_join(size.indx) -> div.df

residfits.c2.w %>%
  left_join(div.df)  -> rich.df
write.csv(rich.df, file = "derived_data/richness_C2_resid.csv")
```


fit models
```{r}
rich.df2 <- read.csv("derived_data/richness_C2_resid.csv", row.names = 1)
head(rich.df)
unique(rich.df$community)

fit.mods.c2 <- function(rich.df2, curr.comm){
  
  rich.df2 %>%
    filter(community == curr.comm) -> df
  mod1 <- lm(w.t50 ~ mean, data = df)
  mod2 <- lm(beta ~ mean, data = df)
  mod3 <- lm(alpha ~ mean, data = df)
  mod4 <- lm(w.r2 ~ mean, data = df)
  mod.list <- list(mod1, mod2, mod3, mod4)
  names(mod.list) <- unlist(code.respVars)
  
  # model r2s
  mod.list %>%
    map(~r2(.x)) %>%
    map_df(.f = unlist, use.names = T, .id = "respvar") %>%
    set_names(nm = c("respvar","Rsq","Rsq.adj")) %>%
    mutate(community = curr.comm) -> df.r2
  df.r2
  # anova table
  mod.list %>%
    map(~aov(.x)) %>%
    map(~model_parameters(.x, eta_squared = "partial")) %>%
    map_df(.f = tibble, .id = "respvar") -> df.an
  # coefs
  mod.list %>%
    map(~model_parameters(.x)) %>%
    map_df(.f = tibble, .id = "respvar") -> df.coefs # coefs

  result.list <- list(r2 = df.r2, an = df.an, coefs = df.coefs)
  return(result.list)
  
}


unique(rich.df2$community) %>%
  set_names() %>%
  map(~fit.mods.c2(rich.df2 = rich.df2, curr.comm = .x)) -> out.c2
out.c2 %>%
  map(~tibble(.x)) %>%
  map(~unnest(.x, cols= c(.x))) %>%
  map_df(.f = tibble, .id = "comm") %>%
  arrange(comm, respvar) -> out.c2.df
out.c2.df

# summarize signif models
out.c2.df %>%
  filter(Parameter %in% c("mean")) %>%
  filter(p < 0.05) -> signif.c2.df
signif.c2.df
out.c2.df %>%
  filter(p<0.05)

```

___________________
# 3-r. STEM-S1. Test role of endophyte diversity after accounting for traits. Exclude C fraction data, include small stems.


calculate diversity metrics and set up dataframe
```{r}
otu.list %>%
  map(~calc.div(.x)) -> div.list

# richness only
div.list %>%
  map(~left_join(.x, seqSamples)) %>%
  list_to_df() %>%
  dplyr::rename('community'='source') %>%
  group_by(community, codeStem) %>%
  summarize(mean=mean(SR),
            se=sd(SR)/sqrt(length(SR)),
            n = sum(!is.na(SR)),
            upper=mean+se,
            lower=mean-se) -> div.df
head(div.df)

resids_explainDecay.list <- readRDS(file = "derived_data/resids_explainDecay_list.RData")
residfits.s1<- resids_explainDecay.list[['S1']]
residfits.s1 %>%
  spread(key = resp, value = resid) -> residfits.s1.w
colnames(residfits.s1.w)

size.indx <- unique(seqSamples[,c("codeStem","size")])
div.df %>%
  left_join(residfits.s1.w) %>%
  left_join(size.indx)-> rich.df
head(rich.df)
write.csv(rich.df, file = "derived_data/richness_S1_resid.csv")
```

fit models
```{r}
rich.df <- read.csv("derived_data/richness_S1_resid.csv", row.names = 1)
head(rich.df)
unique(rich.df$community)

fit.mods.s1 <- function(rich.df, curr.comm){
  
  rich.df %>%
    filter(community == curr.comm) -> df
  mod1 <- lm(time7 ~ size * mean, data = df)
  mod2 <- lm(time13 ~ size * mean, data = df)
  mod3 <- lm(time25 ~ size * mean, data = df)
  mod4 <- lm(time37 ~ size * mean, data = df)
  mod5 <- lm(time59 ~ size * mean, data = df)
  mod.list <- list(mod1, mod2, mod3, mod4, mod5)
  names(mod.list) <- unlist(stem.respVars)
  
  # model r2s
  mod.list %>%
    map(~r2(.x)) %>%
    map_df(.f = unlist, use.names = T, .id = "respvar") %>%
    set_names(nm = c("respvar","Rsq","Rsq.adj")) %>%
    mutate(community = curr.comm) -> df.r2
  df.r2
  # anova table
  mod.list %>%
    map(~aov(.x)) %>%
    map(~model_parameters(.x, eta_squared = "partial")) %>%
    map_df(.f = tibble, .id = "respvar") -> df.an
  # coefs
  mod.list %>%
    map(~model_parameters(.x)) %>%
    map_df(.f = tibble, .id = "respvar") -> df.coefs # coefs
  
  result.list <- list(r2 = df.r2, an = df.an, coefs = df.coefs)
  return(result.list)
  
}

unique(rich.df$community) %>%
  set_names() %>%
  map(~fit.mods.s1(rich.df = rich.df, curr.comm = .x)) -> out.s1
out.s1
out.s1 %>%
  map(~tibble(.x)) %>%
  map(~unnest(.x, cols= c(.x))) %>%
  map_df(.f = tibble, .id = "comm") %>%
  arrange(comm, respvar) -> out.s1.df

out.s1.df
# summarize signif models
out.s1.df %>%
  filter(Parameter %in% c("mean","size:mean")) %>%
  filter(p < 0.05) -> signif.s1.df
signif.s1.df

out.s1.df %>%
  filter(comm == "patho" & respvar == "time37") %>%
  filter(!is.na(p)) %>%
  select(comm, respvar, Parameter, Coefficient, p) %>%
  mutate(p = round(p, digits = 3))

```

examine interactions
```{r}
rich.df %>%
    filter(community == "all") -> df
mod <- lm(time37 ~ size * mean, data = df)
model_parameters(mod) 
estimate_slopes(mod)


```

___________________
# 4-r. STEM-S2. Test role of endophyte diversity after accounting for traits. C fraction data.

calculate diversity metrics and set up dataframe
```{r}
otu.list %>%
  map(~calc.div(.x)) -> div.list

# richness only
div.list %>%
  map(~left_join(.x, seqSamples)) %>%
  list_to_df() %>%
  dplyr::rename('community'='source') %>%
  group_by(community, codeStem) %>%
  summarize(mean=mean(SR),
            se=sd(SR)/sqrt(length(SR)),
            n = sum(!is.na(SR)),
            upper=mean+se,
            lower=mean-se) -> div.df
head(div.df)

resids_explainDecay.list <- readRDS(file = "derived_data/resids_explainDecay_list.RData")
residfits.s2<- resids_explainDecay.list[['S2']]
residfits.s2 %>%
  spread(key = resp, value = resid) -> residfits.s2.w
colnames(residfits.s2.w)

size.indx <- unique(seqSamples[,c("codeStem","size")])
div.df %>%
  left_join(size.indx)-> tmp
tmp
tmp %>%
  left_join(residfits.s2.w) -> rich.df
head(rich.df)
write.csv(rich.df, file = "derived_data/richness_S2_resid.csv")
```

fit models
```{r}
rich.df2 <- read.csv("derived_data/richness_S2_resid.csv", row.names = 1)
head(rich.df)
unique(rich.df$community)
fit.mods.s2 <- function(rich.df2, curr.comm){
  
  rich.df2 %>%
    filter(community == curr.comm) -> df
  mod1 <- lm(time7 ~ mean, data = df)
  mod2 <- lm(time13 ~ mean, data = df)
  mod3 <- lm(time25 ~ mean, data = df)
  mod4 <- lm(time37 ~ mean, data = df)
  mod5 <- lm(time59 ~ mean, data = df)
  mod.list <- list(mod1, mod2, mod3, mod4, mod5)
  names(mod.list) <- unlist(stem.respVars)
  
  # model r2s
  mod.list %>%
    map(~r2(.x)) %>%
    map_df(.f = unlist, use.names = T, .id = "respvar") %>%
    set_names(nm = c("respvar","Rsq","Rsq.adj")) %>%
    mutate(community = curr.comm) -> df.r2
  df.r2
  # anova table
  mod.list %>%
    map(~aov(.x)) %>%
    map(~model_parameters(.x, eta_squared = "partial")) %>%
    map_df(.f = tibble, .id = "respvar") -> df.an
  # coefs
  mod.list %>%
    map(~model_parameters(.x)) %>%
    map_df(.f = tibble, .id = "respvar") -> df.coefs # coefs
  
  result.list <- list(r2 = df.r2, an = df.an, coefs = df.coefs)
  return(result.list)
  
}

unique(rich.df2$community) %>%
  set_names() %>%
  map(~fit.mods.s2(rich.df2 = rich.df2, curr.comm = .x)) -> out.s2
out.s2 %>%
  map(~tibble(.x)) %>%
  map(~unnest(.x, cols= c(.x))) %>%
  map_df(.f = tibble, .id = "comm") %>%
  arrange(comm, respvar) -> out.s2.df

# summarize signif models
out.s2.df %>%
  filter(Parameter %in% c("mean","size:mean")) %>%
  filter(p < 0.05) -> signif.s2.df
signif.s2.df

out.s2.df %>%
  filter(comm == "oo" & respvar == "time25") %>%
  filter(!is.na(p)) %>%
  select(comm, respvar, Parameter, Coefficient, p) %>%
  mutate(p = round(p, digits = 3))



```
